Проверка при удалении системной роли

+++ auth_api/src_auth/features/roles/v1/service.py
@@ -3,7 +3,7 @@ from uuid import UUID
 
 from fastapi import Depends
 
-from src_auth.core.exc.exceptions import RoleNotFoundError, UserOrRoleNotFoundError
+from src_auth.core.exc.exceptions import RoleNotFoundError, UserOrRoleNotFoundError, AccessDeniedError
 from src_auth.features.roles.v1.dto import CreateRoleDTO, RoleDTO
 from src_auth.features.roles.v1.repository import (
     RoleRepositoryInterface,
@@ -43,6 +43,9 @@ class RoleService:
         name: str | None = None,
         description: str | None = None,
     ) -> RoleDTO:
+        if await self.repository.is_system_role(role_id=role_id):
+            raise AccessDeniedError("Cannot update a system role")
+        
         updated = await self.repository.update_role(
             role_id=role_id,
             name=name,
@@ -57,6 +60,9 @@ class RoleService:
         self,
         role_id: UUID,
     ) -> bool:
+        if await self.repository.is_system_role(role_id=role_id):
+            raise AccessDeniedError("Cannot delete system role")
+        
         is_deleted = await self.repository.delete_role(role_id=role_id)
         if not is_deleted:
             raise RoleNotFoundError("Role not found")
